update map marker faster
